tests/wpt/css-tests/css-shapes-1_dev/xhtml1/shape-outside-invalid-circle-001.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
    <meta charset="utf-8" />
    <title>CSS Shapes Tests: Circle - invalid r values</title>
    <link href="biqing.kwok@qq.com" rel="author" title="biqing" />
    <link href="http://www.w3.org/TR/css-shapes-1/#supported-basic-shapes" rel="help" />
    <link href="http://www.w3.org/TR/css-shapes-1/#funcdef-circle" rel="help" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link href="/resources/testharness.css" rel="stylesheet" />
    <meta content="dom" name="flags" />
    <meta content="shape-outside declaration is invalid if circle shape function has an invalid r parameter" name="assert" />
    <style type="text/css">
    .circle {
        float: left;
        width: 100px;
        height: 100px;
    }
    #circle1 {
        shape-outside: circle(auto);
    }
    #circle2 {
        shape-outside: circle(inherit);
    }
    #circle3 {
        shape-outside: circle(#FFFFFF);
    }
    </style>
</head>
<body>
    <div id="log"></div>

    <div id="circle1" class="circle"></div>
    <div id="circle2" class="circle"></div>
    <div id="circle3" class="circle"></div>

    <script>

    test(function() {
        assert_true(getComputedStyle(document.getElementById('circle1')).shapeOutside == 'none');
        assert_true(getComputedStyle(document.getElementById('circle2')).shapeOutside == 'none');
        assert_true(getComputedStyle(document.getElementById('circle3')).shapeOutside == 'none');
    }, 'Circle - invalid r values');

    </script>

</body></html>
